fix: rename queries/transaction directory to queries/transactions
