+++ src/pages/statistics/statistics.tsx
@@ -2,7 +2,7 @@ import { styled } from 'styled-components'
 import { DonutChart } from './components/donut-chart'
 import { Flex, Section, Text } from '@radix-ui/themes'
 import { transactionAmountFormat } from '~/utils/units'
-import { Transaction, useTransactionsListQuery } from '~/queries/transaction'
+import { Transaction, useTransactionsListQuery } from '~/queries/transactions'
 import { toNumber } from '~/utils/number'
 import { SORT_ORDER } from '~/constants/query'
 import { useMemo } from 'react'
