+++ src/pages/transaction-create/components/optional-info-fields.type.ts
@@ -1,7 +1,7 @@
 import { Dispatch } from 'react'
 import { OptionalInfoAction } from './transaction-form.type'
 
-import { OptionalInfo } from '~/queries/transaction/transaction.type'
+import { OptionalInfo } from '~/queries/transactions/transactions.type'
 
 export interface OptionalInfoProps {
   optionalInfo: OptionalInfo
