+++ src/pages/transaction-create/components/required-info-fields.type.ts
@@ -1,7 +1,7 @@
 import { Dispatch } from 'react'
 
 import { RequiredInfoAction } from './transaction-form.type'
-import { RequiredInfo } from '~/queries/transaction/transaction.type'
+import { RequiredInfo } from '~/queries/transactions/transactions.type'
 
 export interface RequiredInfoProps {
   requiredInfo: RequiredInfo
