--- src/queries/transaction/index.ts
@@ -1,2 +1,0 @@
-export * from './transaction.api'
-export * from './transaction.type'
